Split account settings sections
