+++ apps/web/src/components/settings/AccountSettings.tsx
@@ -1,581 +1,15 @@
 "use client";
 
-import { useState, useEffect, useCallback, useRef } from "react";
-import { useRouter } from "next/navigation";
-import { Trans } from "@lingui/react/macro";
-import { useLingui } from "@lingui/react/macro";
-import { GitHubIcon } from "@/components/icons/GitHubIcon";
-import { authClient } from "@/lib/auth-client";
-import { isReservedUsername } from "@/lib/username";
+import { useCallback, useState } from "react";
 import { useSession } from "@/components/providers/SessionProvider";
-import { useLocalePath } from "@/lib/useLocalePath";
-import {
-  setPassword as setPasswordAction,
-  recordPasswordResetRequest,
-  getAccountPageData,
-  renameUsername,
-} from "@/lib/actions/preferences";
-import { Button } from "@/components/ui/Button";
-import { FormField } from "@/components/ui/FormField";
-import { ErrorAlert } from "@/components/ui/ErrorAlert";
-import { SuccessAlert } from "@/components/ui/SuccessAlert";
-import { GoogleIcon } from "@/components/icons/GoogleIcon";
-import { LinkedInIcon } from "@/components/icons/LinkedInIcon";
-
-/* ── Types ── */
-
-type ConnectedAccount = {
-  providerId: string;
-  accountId: string;
-};
-
-/* ── Login prompt for unauthenticated users ── */
-
-function LoginPrompt() {
-  const { t } = useLingui();
-  const lp = useLocalePath();
-  return (
-    <div className="flex flex-col items-center gap-4 py-12 text-center">
-      <p className="text-muted">
-        <Trans id="settings.account.loginRequired" comment="Message when user must log in to see account settings">
-          Please log in to manage your account settings.
-        </Trans>
-      </p>
-      <Button href={lp("/sign-in")} variant="primary" size="md">
-        {t({ id: "common.auth.login", comment: "Login button label", message: "Log in" })}
-      </Button>
-    </div>
-  );
-}
-
-/* ── Password Section ── */
-
-function PasswordSection({ hasPassword, initialCooldown, onPasswordSet }: { hasPassword: boolean; initialCooldown: number; onPasswordSet: () => void }) {
-  if (hasPassword) return <ResetPasswordFlow initialCooldown={initialCooldown} />;
-  return <SetPasswordFlow onSuccess={onPasswordSet} />;
-}
-
-function SetPasswordFlow({ onSuccess }: { onSuccess: () => void }) {
-  const { t } = useLingui();
-  const { user } = useSession();
-  const [newPassword, setNewPassword] = useState("");
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState("");
-  const [newPasswordError, setNewPasswordError] = useState("");
-  const [success, setSuccess] = useState("");
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    setError("");
-    setNewPasswordError("");
-    setSuccess("");
-    if (!newPassword) {
-      const message = t({ id: "settings.account.password.newRequired", comment: "Error when new password is empty", message: "Please enter a password" });
-      setNewPasswordError(message);
-      setError(message);
-      return;
-    }
-    setLoading(true);
-    const result = await setPasswordAction(newPassword);
-    setLoading(false);
-    if (result.error) {
-      setNewPasswordError("");
-      setError(result.error ?? t({ id: "settings.account.password.setError", comment: "Generic set password error", message: "Failed to set password" }));
-    } else {
-      setSuccess(t({ id: "settings.account.password.setSuccess", comment: "Success message after setting password", message: "Password set successfully." }));
-      setNewPassword("");
-      setTimeout(onSuccess, 2000);
-    }
-  }
-
-  return (
-    <section>
-      <h2 className="mb-1 text-base font-semibold">
-        <Trans id="settings.account.password.title" comment="Password section heading">Password</Trans>
-      </h2>
-      <p className="mb-4 text-sm text-muted">
-        <Trans id="settings.account.password.setDescription" comment="Set password description for OAuth users">
-          You signed in with a social account. Set a password to enable email changes and additional security.
-        </Trans>
-      </p>
-      <ErrorAlert message={error} focusOnRender />
-      <SuccessAlert message={success} />
-      <form onSubmit={handleSubmit} className="flex flex-col gap-4 min-[480px]:flex-row min-[480px]:items-end">
-        <input type="hidden" name="username" autoComplete="username" value={user?.email ?? ""} />
-        <div className="flex-1">
-          <FormField
-            label={t({ id: "settings.account.password.newLabel", comment: "New password input label", message: "New password" })}
-            type="password"
-            required
-            autoComplete="new-password"
-            value={newPassword}
-            onChange={(e) => {
-              setNewPassword(e.target.value);
-              setNewPasswordError("");
-            }}
-            error={newPasswordError}
-          />
-        </div>
-        <Button type="submit" disabled={loading} size="sm">
-          {loading
-            ? t({ id: "settings.account.password.setting", comment: "Set password button while loading", message: "Setting..." })
-            : t({ id: "settings.account.password.set", comment: "Set password button", message: "Set password" })}
-        </Button>
-      </form>
-    </section>
-  );
-}
-
-function ResetPasswordFlow({ initialCooldown }: { initialCooldown: number }) {
-  const { t } = useLingui();
-  const { user } = useSession();
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState("");
-  const [success, setSuccess] = useState("");
-  const [cooldown, setCooldown] = useState(initialCooldown);
-
-  useEffect(() => {
-    if (cooldown <= 0) return;
-    const timer = setInterval(() => {
-      setCooldown((prev) => {
-        const next = prev - 1;
-        if (next <= 0) clearInterval(timer);
-        return Math.max(0, next);
-      });
-    }, 1000);
-    return () => clearInterval(timer);
-  }, [cooldown]);
-
-  async function handleReset() {
-    if (!user?.email) return;
-    setError("");
-    setSuccess("");
-    setLoading(true);
-
-    const result = await recordPasswordResetRequest();
-    if (result.cooldown) {
-      setCooldown(result.cooldown);
-      setLoading(false);
-      return;
-    }
-    if (result.error) {
-      setError(result.error);
-      setLoading(false);
-      return;
-    }
-
-    const { error } = await authClient.requestPasswordReset({
-      email: user.email,
-      redirectTo: "/reset-password",
-    });
-    setLoading(false);
-    if (error) {
-      setError(error.message ?? t({ id: "settings.account.password.error", comment: "Generic password reset error", message: "Failed to send reset email" }));
-    } else {
-      setSuccess(t({ id: "settings.account.password.success", comment: "Success message after password reset request", message: "Password reset link sent. Please check your inbox. It may take a few minutes to arrive." }));
-      setCooldown(60);
-    }
-  }
-
-  return (
-    <section>
-      <h2 className="mb-1 text-base font-semibold">
-        <Trans id="settings.account.password.title" comment="Password section heading">Password</Trans>
-      </h2>
-      <p className="mb-4 text-sm text-muted">
-        <Trans id="settings.account.password.description" comment="Password section description">
-          Change your password via a secure email link.
-        </Trans>
-      </p>
-      <ErrorAlert message={error} focusOnRender />
-      <SuccessAlert message={success} />
-      <Button onClick={handleReset} disabled={loading || cooldown > 0} size="sm">
-        {loading
-          ? t({ id: "settings.account.password.sending", comment: "Reset password button while loading", message: "Sending..." })
-          : cooldown > 0
-            ? t({ id: "settings.account.password.cooldown", comment: "Reset password button during cooldown with seconds remaining", message: `Resend in ${cooldown}s` })
-            : t({ id: "settings.account.password.send", comment: "Reset password button", message: "Send reset link" })}
-      </Button>
-    </section>
-  );
-}
-
-/* ── Username ── */
-
-const USERNAME_RE = /^[a-z0-9][a-z0-9-]*[a-z0-9]$/;
-
-function UsernameSection({ currentUsername }: { currentUsername: string }) {
-  const { t } = useLingui();
-  const router = useRouter();
-  const { refresh: refreshSession } = useSession();
-  const [savedUsername, setSavedUsername] = useState(currentUsername);
-  const [value, setValue] = useState(currentUsername);
-  const [loading, setLoading] = useState(false);
-  const [checking, setChecking] = useState(false);
-  const [available, setAvailable] = useState<boolean | null>(null);
-  const [error, setError] = useState("");
-  const [success, setSuccess] = useState("");
-  const debounceRef = useRef<ReturnType<typeof setTimeout>>(null);
-
-  useEffect(() => {
-    return () => {
-      if (debounceRef.current) clearTimeout(debounceRef.current);
-    };
-  }, []);
-
-  const normalized = value.toLowerCase().trim();
-  const unchanged = normalized === savedUsername;
-  const tooShort = normalized.length < 3;
-  const tooLong = normalized.length > 30;
-  const invalidChars = normalized.length >= 3 && !USERNAME_RE.test(normalized);
-  const reserved = !invalidChars && normalized.length >= 3 && isReservedUsername(normalized);
-
-  function handleChange(raw: string) {
-    const v = raw.toLowerCase().replace(/[^a-z0-9-]/g, "");
-    setValue(v);
-    setError("");
-    setSuccess("");
-    setAvailable(null);
-
-    if (debounceRef.current) clearTimeout(debounceRef.current);
-
-    const norm = v.trim();
-    if (norm === savedUsername || norm.length < 3 || norm.length > 30 || !USERNAME_RE.test(norm) || isReservedUsername(norm)) return;
-
-    setChecking(true);
-    debounceRef.current = setTimeout(async () => {
-      try {
-        const res = await authClient.isUsernameAvailable({ username: norm });
-        setAvailable(!!res.data?.available);
-      } catch {
-        setAvailable(null);
-      }
-      setChecking(false);
-    }, 400);
-  }
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    if (unchanged || tooShort || tooLong || invalidChars || reserved || available === false) return;
-    setError("");
-    setSuccess("");
-    setLoading(true);
-    // Server action wraps `auth.api.updateUser` AND fans out every
-    // cache layer the rename invalidates (Redis session, watchlist
-    // cache tags, Typesense `owner_username`, sitemap). See #3022 +
-    // the action's docstring.
-    try {
-      const { error } = await renameUsername(normalized);
-      if (error) {
-        setLoading(false);
-        setError(error);
-        return;
-      }
-    } catch {
-      setLoading(false);
-      setError(t({ id: "settings.account.username.error", comment: "Generic username update error", message: "Failed to update username" }));
-      return;
-    }
-    // Re-pull the bootstrap payload so SessionProvider stops handing
-    // the stale `user.username` to URL-building components like
-    // WatchlistCard / save-search / mirror. router.refresh() rebuilds
-    // any RSC tree currently rendered with the old slug.
-    await refreshSession();
-    router.refresh();
-    setLoading(false);
-    setSavedUsername(normalized);
-    setAvailable(null);
-    setSuccess(t({ id: "settings.account.username.success", comment: "Username updated success message", message: "Username updated." }));
-  }
-
-  // hint is derived from the validated flags above (which already use savedUsername)
-  const hint = unchanged
-    ? null
-    : tooShort
-      ? t({ id: "settings.account.username.tooShort", comment: "Username too short hint", message: "At least 3 characters" })
-      : tooLong
-        ? t({ id: "settings.account.username.tooLong", comment: "Username too long hint", message: "At most 30 characters" })
-        : invalidChars
-          ? t({ id: "settings.account.username.invalidChars", comment: "Username invalid characters hint", message: "Only lowercase letters, numbers, and hyphens (cannot start/end with hyphen)" })
-          : reserved
-            ? t({ id: "settings.account.username.reserved", comment: "Username is reserved hint", message: "This username is reserved" })
-            : checking
-            ? t({ id: "settings.account.username.checking", comment: "Checking username availability", message: "Checking availability..." })
-            : available === true
-              ? t({ id: "settings.account.username.available", comment: "Username is available", message: "Available" })
-              : available === false
-                ? t({ id: "settings.account.username.taken", comment: "Username is taken", message: "Already taken" })
-                : null;
-  const usernameValidationInvalid = !unchanged && (tooShort || tooLong || invalidChars || reserved || available === false);
-  const usernameInvalid = usernameValidationInvalid || !!error;
-  const hintClassName = available === true
-    ? "mt-1 text-xs text-green-600 dark:text-green-400"
-    : usernameInvalid
-      ? "mt-1 text-xs text-error"
-      : "mt-1 text-xs text-muted";
-
-  return (
-    <section>
-      <h2 className="mb-1 text-base font-semibold">
-        <Trans id="settings.account.username.title" comment="Username section heading">Username</Trans>
-      </h2>
-      <p className="mb-4 text-sm text-muted">
-        <Trans id="settings.account.username.description" comment="Username section description">
-          Your unique handle used in your public profile URL.
-        </Trans>
-      </p>
-      <ErrorAlert message={error} focusOnRender />
-      <form onSubmit={handleSubmit}>
-        <div className="flex flex-col gap-4 min-[480px]:flex-row min-[480px]:items-end">
-          <div className="flex-1">
-            <FormField
-              label={t({ id: "settings.account.username.label", comment: "Username input label", message: "Username" })}
-              required
-              autoComplete="username"
-              value={value}
-              onChange={(e) => handleChange(e.target.value)}
-              maxLength={30}
-              hint={hint}
-              hintClassName={hintClassName}
-              error={error}
-              aria-invalid={usernameInvalid || undefined}
-            />
-          </div>
-          <Button type="submit" disabled={loading || unchanged || tooShort || tooLong || invalidChars || reserved || available === false || checking} size="sm">
-            {loading
-              ? t({ id: "settings.account.username.saving", comment: "Username save button while loading", message: "Saving..." })
-              : t({ id: "settings.account.username.save", comment: "Username save button", message: "Update username" })}
-          </Button>
-        </div>
-        {success && <div className="mt-3"><SuccessAlert message={success} /></div>}
-      </form>
-    </section>
-  );
-}
-
-/* ── Change Email ── */
-
-function ChangeEmailSection() {
-  const { t } = useLingui();
-  const lp = useLocalePath();
-  const [newEmail, setNewEmail] = useState("");
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState("");
-  const [emailError, setEmailError] = useState("");
-  const [success, setSuccess] = useState("");
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    setError("");
-    setEmailError("");
-    setSuccess("");
-    if (!newEmail) {
-      const message = t({ id: "settings.account.email.required", comment: "Error when email field is empty", message: "Please enter a new email address" });
-      setEmailError(message);
-      setError(message);
-      return;
-    }
-    setLoading(true);
-    const { error } = await authClient.changeEmail({
-      newEmail,
-      callbackURL: lp("/settings/account"),
-    });
-    setLoading(false);
-    if (error) {
-      setEmailError("");
-      setError(error.message ?? t({ id: "settings.account.email.error", comment: "Generic email change error", message: "Failed to change email" }));
-    } else {
-      setSuccess(t({ id: "settings.account.email.success", comment: "Success message after email change request", message: "Verification email sent. Please check your inbox. It may take a few minutes to arrive." }));
-      setNewEmail("");
-    }
-  }
-
-  return (
-    <section>
-      <h2 className="mb-1 text-base font-semibold">
-        <Trans id="settings.account.email.title" comment="Change email section heading">Change email</Trans>
-      </h2>
-      <p className="mb-4 text-sm text-muted">
-        <Trans id="settings.account.email.description" comment="Change email section description">
-          Update the email address associated with your account.
-        </Trans>
-      </p>
-      <ErrorAlert message={error} focusOnRender />
-      <SuccessAlert message={success} />
-      <form onSubmit={handleSubmit} className="flex flex-col gap-4 min-[480px]:flex-row min-[480px]:items-end">
-        <div className="flex-1">
-          <FormField
-            label={t({ id: "settings.account.email.label", comment: "New email input label", message: "New email" })}
-            type="email"
-            required
-            autoComplete="email"
-            value={newEmail}
-            onChange={(e) => {
-              setNewEmail(e.target.value);
-              setEmailError("");
-            }}
-            error={emailError}
-          />
-        </div>
-        <Button type="submit" disabled={loading} size="sm">
-          {loading
-            ? t({ id: "settings.account.email.saving", comment: "Email save button while loading", message: "Saving..." })
-            : t({ id: "settings.account.email.save", comment: "Email save button", message: "Update email" })}
-        </Button>
-      </form>
-    </section>
-  );
-}
-
-/* ── Connected Accounts ── */
-
-const socialProviders = [
-  { id: "github", label: "GitHub", icon: <GitHubIcon size={20} /> },
-  { id: "google", label: "Google", icon: <GoogleIcon size={20} /> },
-  { id: "linkedin", label: "LinkedIn", icon: <LinkedInIcon size={20} /> },
-] as const;
-
-function ConnectedAccountsSection({ accounts, onDisconnect }: { accounts: ConnectedAccount[]; onDisconnect: (providerId: string) => void }) {
-  const { t } = useLingui();
-  const lp = useLocalePath();
-  const [actionLoading, setActionLoading] = useState<string | null>(null);
-  const [error, setError] = useState("");
-
-  function isConnected(providerId: string) {
-    return accounts.some((a) => a.providerId === providerId);
-  }
-
-  async function handleConnect(provider: string) {
-    setActionLoading(provider);
-    setError("");
-    const result = await authClient.linkSocial({
-      provider: provider as "github" | "google" | "linkedin",
-      callbackURL: lp("/settings/account"),
-    });
-    if (result.error) {
-      setError(result.error.message ?? t({ id: "settings.account.socials.connectError", comment: "Error when linking social account fails", message: "Failed to connect account" }));
-    }
-    setActionLoading(null);
-  }
-
-  async function handleDisconnect(providerId: string) {
-    setActionLoading(providerId);
-    setError("");
-    try {
-      const result = await authClient.unlinkAccount({ providerId });
-      if (result.error) {
-        setError(result.error.message ?? t({ id: "settings.account.socials.disconnectError", comment: "Error when unlinking social account fails", message: "Failed to disconnect account" }));
-      } else {
-        onDisconnect(providerId);
-      }
-    } catch {
-      setError(t({ id: "settings.account.socials.disconnectError", comment: "Error when unlinking social account fails", message: "Failed to disconnect account" }));
-    }
-    setActionLoading(null);
-  }
-
-  return (
-    <section>
-      <h2 className="mb-1 text-base font-semibold">
-        <Trans id="settings.account.socials.title" comment="Connected accounts section heading">Connected accounts</Trans>
-      </h2>
-      <p className="mb-4 text-sm text-muted">
-        <Trans id="settings.account.socials.description" comment="Connected accounts section description">
-          Manage your linked social accounts.
-        </Trans>
-      </p>
-      <ErrorAlert message={error} focusOnRender />
-      <div className="space-y-2">
-        {socialProviders.map((p) => {
-          const connected = isConnected(p.id);
-          return (
-            <div key={p.id} className="flex items-center justify-between rounded-md border border-divider px-4 py-3">
-              <div className="flex items-center gap-3">
-                {p.icon}
-                <span className="text-sm font-medium">{p.label}</span>
-              </div>
-              <Button
-                onClick={() => (connected ? handleDisconnect(p.id) : handleConnect(p.id))}
-                disabled={actionLoading === p.id}
-                variant={connected ? "outline" : "primary"}
-                size="sm"
-              >
-                {connected
-                  ? t({ id: "settings.account.socials.disconnect", comment: "Disconnect social account button", message: "Disconnect" })
-                  : t({ id: "settings.account.socials.connect", comment: "Connect social account button", message: "Connect" })}
-              </Button>
-            </div>
-          );
-        })}
-      </div>
-    </section>
-  );
-}
-
-/* ── Delete Account ── */
-
-function DeleteAccountSection() {
-  const { t } = useLingui();
-  const [showConfirm, setShowConfirm] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState("");
-
-  async function handleDelete() {
-    setError("");
-    setLoading(true);
-    const { error } = await authClient.deleteUser({});
-    setLoading(false);
-    if (error) {
-      setError(error.message ?? t({ id: "settings.account.delete.error", comment: "Generic account deletion error", message: "Failed to delete account" }));
-    } else {
-      window.location.href = "/";
-    }
-  }
-
-  return (
-    <section className="rounded-md border border-error-border bg-error-bg p-4">
-      <h2 className="mb-1 text-base font-semibold text-error">
-        <Trans id="settings.account.delete.title" comment="Delete account section heading">Delete account</Trans>
-      </h2>
-      <p className="mb-4 text-sm text-error">
-        <Trans id="settings.account.delete.description" comment="Delete account section description">
-          Permanently delete your account and all associated data. This action cannot be undone.
-        </Trans>
-      </p>
-      <ErrorAlert message={error} focusOnRender />
-      {!showConfirm ? (
-        <Button onClick={() => setShowConfirm(true)} variant="danger" size="sm">
-          {t({ id: "settings.account.delete.button", comment: "Delete account button", message: "Delete my account" })}
-        </Button>
-      ) : (
-        <div className="flex gap-2">
-          <Button onClick={handleDelete} disabled={loading} variant="danger" size="sm">
-            {loading
-              ? t({ id: "settings.account.delete.deleting", comment: "Delete button while loading", message: "Deleting..." })
-              : t({ id: "settings.account.delete.confirm", comment: "Confirm delete button", message: "Confirm deletion" })}
-          </Button>
-          <Button
-            onClick={() => { setShowConfirm(false); setError(""); }}
-            variant="danger-outline"
-            size="sm"
-          >
-            {t({ id: "settings.account.delete.cancel", comment: "Cancel delete button", message: "Cancel" })}
-          </Button>
-        </div>
-      )}
-    </section>
-  );
-}
-
-/* ── Types ── */
-
-type AccountPageData = {
-  accounts: ConnectedAccount[];
-  hasPassword: boolean;
-  username: string;
-} | null;
-
-/* ── Main Component ── */
+import { getAccountPageData } from "@/lib/actions/preferences";
+import { ChangeEmailSection } from "./account/ChangeEmailSection";
+import { ConnectedAccountsSection } from "./account/ConnectedAccountsSection";
+import { DeleteAccountSection } from "./account/DeleteAccountSection";
+import { LoginPrompt } from "./account/LoginPrompt";
+import { PasswordSection } from "./account/PasswordSection";
+import { UsernameSection } from "./account/UsernameSection";
+import type { AccountPageData, ConnectedAccount } from "./account/types";
 
 export function AccountSettings({ initialData }: { initialData?: AccountPageData }) {
   const { isLoggedIn } = useSession();
