+++ apps/web/src/components/settings/account/types.ts
@@ -1,0 +1,10 @@
+export type ConnectedAccount = {
+  providerId: string;
+  accountId: string;
+};
+
+export type AccountPageData = {
+  accounts: ConnectedAccount[];
+  hasPassword: boolean;
+  username: string;
+} | null;
